requirements.txt
openai==0.28.0
aiogram==2.0
psycopg2-binary==2.9.9
python-dotenv==0.21.0